incremento de resultado da expressao caso correta

+++ n1/primeiroCompilador/token.h
@@ -2,6 +2,7 @@
 #define TOKEN_H
 
 #include "enums.h"
+#include <stdio.h>
 
 namespace COMPILADOR{
 
@@ -18,14 +19,14 @@ public:
     Token(ENUMS::tipoToken tipo, double valor);
 
     //gets
-    ENUMS::tipoToken getTipo() const    {return this->tipo;}
-    int getValorInt() const                {return this->valorInt;}
-    int getValorDouble() const             {return this->valorDouble;}
+    ENUMS::tipoToken getTipo() const        {return this->tipo;}
+    int getValorInt() const                 {return this->valorInt;}
+    double getValorDouble() const           {return this->valorDouble;}
 
     //sets
-    void setTipo(ENUMS::tipoToken tipo) {this->tipo = tipo;}
-    void setValorInt(int valor)            {this->valorInt = valor;}
-    void setValorDouble(int valor)            {this->valorDouble = valor;}
+    void setTipo(ENUMS::tipoToken tipo)     {this->tipo = tipo;}
+    void setValorInt(int valor)             {this->valorInt = valor;}
+    void setValorDouble(double valor)       {this->valorDouble = valor;}
 };
